@@ -1,13 +1,15 @@
 ---
 name: travelcrew
 description: Travel Crew is a mobile app for friends to share travel plans, organize travel and even create full itineraries together.
-version: 3.0.8+72
+version: 3.0.9+73
 publish_to: none
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 dependencies:
   flutter:
+    sdk: flutter
+  flutter_localizations:
     sdk: flutter
 
   # API
@@ -133,3 +135,7 @@ flutter:
     - asset: assets/fonts/PlayfairDisplay-VariableFont_wght.ttf
     - asset: assets/fonts/PlayfairDisplay-Italic-VariableFont_wght.ttf
       style: italic
+
+
+flutter_intl:
+  enabled: true
