@@ -10,8 +10,8 @@ import 'current_profile_state.dart';
 
 
 class CurrentProfileBloc extends Bloc<CurrentProfileEvent, CurrentProfileState> {
-  final CurrentUserProfileRepository currentUserProfileRepository;
-  StreamSubscription _subscription;
+  final CurrentUserProfileRepository? currentUserProfileRepository;
+  StreamSubscription? _subscription;
 
 
   CurrentProfileBloc({this.currentUserProfileRepository}) : super(CurrentProfileLoadingState());
@@ -24,16 +24,16 @@ class CurrentProfileBloc extends Bloc<CurrentProfileEvent, CurrentProfileState> 
       if(_subscription != null){
         await _subscription?.cancel();
       }
-      _subscription = currentUserProfileRepository.profile().asBroadcastStream().listen((data) { add(CurrentProfileHasDataEvent(data)); });
+      _subscription = currentUserProfileRepository?.profile().asBroadcastStream().listen((data) { add(CurrentProfileHasDataEvent(data)); });
     }
     else if(event is CurrentProfileHasDataEvent){
       yield CurrentProfileHasDataState(event.data);
     }
   }
   @override
-  Future<Function> close(){
+  Future<void> close(){
     _subscription?.cancel();
-    currentUserProfileRepository.dispose();
+    currentUserProfileRepository?.dispose();
     return super.close();
   }
 }
