@@ -1,11 +1,11 @@
 class LoginState {
-  final bool isEmailValid;
-  final bool isPasswordValid;
-  final bool isSubmitting;
-  final bool isSuccess;
-  final bool isFailure;
+  final bool? isEmailValid;
+  final bool? isPasswordValid;
+  final bool? isSubmitting;
+  final bool? isSuccess;
+  final bool? isFailure;
 
-  bool get isFormValid => isEmailValid && isPasswordValid;
+  bool get isFormValid => isEmailValid! && isPasswordValid!;
 
   LoginState(
       {this.isEmailValid,
@@ -55,12 +55,12 @@ class LoginState {
   }
 
   LoginState update({
-    bool isEmailValid,
-    bool isPasswordValid,
+    bool? isEmailValid,
+    bool? isPasswordValid,
   }) {
     return copyWith(
-      isEmailValid: isEmailValid,
-      isPasswordValid: isPasswordValid,
+      isEmailValid: isEmailValid!,
+      isPasswordValid: isPasswordValid!,
       isSubmitting: false,
       isSuccess: false,
       isFailure: false,
@@ -68,11 +68,11 @@ class LoginState {
   }
 
   LoginState copyWith({
-    bool isEmailValid,
-    bool isPasswordValid,
-    bool isSubmitting,
-    bool isSuccess,
-    bool isFailure,
+    bool? isEmailValid,
+    bool? isPasswordValid,
+    bool? isSubmitting,
+    bool? isSuccess,
+    bool? isFailure,
   }) {
     return LoginState(
       isEmailValid: isEmailValid ?? this.isEmailValid,
