@@ -12,7 +12,7 @@ abstract class SignupEvent extends Equatable {
 class SignupFirstNameChanged extends SignupEvent {
   final String firstName;
 
-  SignupFirstNameChanged({this.firstName});
+  SignupFirstNameChanged({required this.firstName});
 
   @override
   List<Object> get props => [firstName];
@@ -20,7 +20,7 @@ class SignupFirstNameChanged extends SignupEvent {
 class SignupLastNameChanged extends SignupEvent {
   final String lastName;
 
-  SignupLastNameChanged({this.lastName});
+  SignupLastNameChanged({required this.lastName});
 
   @override
   List<Object> get props => [lastName];
@@ -28,7 +28,7 @@ class SignupLastNameChanged extends SignupEvent {
 class SignupDisplayNameChanged extends SignupEvent {
   final String displayName;
 
-  SignupDisplayNameChanged({this.displayName});
+  SignupDisplayNameChanged({required this.displayName});
 
   @override
   List<Object> get props => [displayName];
@@ -37,7 +37,7 @@ class SignupDisplayNameChanged extends SignupEvent {
 class SignupImageChanged extends SignupEvent {
   final File urlToImage;
 
-  SignupImageChanged({this.urlToImage});
+  SignupImageChanged({ required this.urlToImage});
 
   @override
   List<Object> get props => [urlToImage];
@@ -45,7 +45,7 @@ class SignupImageChanged extends SignupEvent {
 class SignupEmailChanged extends SignupEvent {
   final String email;
 
-  SignupEmailChanged({this.email});
+  SignupEmailChanged({required this.email});
 
   @override
   List<Object> get props => [email];
@@ -54,7 +54,7 @@ class SignupEmailChanged extends SignupEvent {
 class SignupPasswordChanged extends SignupEvent {
   final String password;
 
-  SignupPasswordChanged({this.password});
+  SignupPasswordChanged({required this.password});
 
   @override
   List<Object> get props => [password];
@@ -63,13 +63,13 @@ class SignupPasswordChanged extends SignupEvent {
 class SignupSubmitted extends SignupEvent {
   final String email;
   final String password;
-  final String firstName;
-  final String lastName;
-  final String displayName;
-  final File urlToImage;
+  final String? firstName;
+  final String? lastName;
+  final String? displayName;
+  final File? urlToImage;
 
-  SignupSubmitted({this.firstName, this.lastName, this.displayName, this.urlToImage, this.email, this.password});
+  SignupSubmitted({this.firstName, this.lastName, this.displayName, this.urlToImage, required this.email, required this.password});
 
   @override
-  List<Object> get props => [email, password, displayName, firstName, lastName, urlToImage];
+  List<Object> get props => [email, password, displayName as Object, firstName as Object, lastName as Object, urlToImage as Object];
 }
