@@ -1,21 +1,21 @@
 
 ///Model for current user
 class CurrentUser {
-  final List<String> blockedList;
-  final String displayName;
-  final String email;
-  final String firstName;
-  final List<String> following;
-  final List<String> followers;
-  final String lastName;
+  final List<String>? blockedList;
+  final String? displayName;
+  final String? email;
+  final String? firstName;
+  final List<String>? following;
+  final List<String>? followers;
+  final String? lastName;
   // final int tripsCreated;
   // final int tripsJoined;
-  final String uid;
-  final String urlToImage;
-  final String hometown;
-  final String instagramLink;
-  final String facebookLink;
-  final List<String> topDestinations;
+  final String? uid;
+  final String? urlToImage;
+  final String? hometown;
+  final String? instagramLink;
+  final String? facebookLink;
+  final List<String>? topDestinations;
   
   CurrentUser({this.blockedList, this.displayName, this.email, this.firstName, this.following, this.followers, this.lastName, this.uid, this.urlToImage, this.hometown, this.instagramLink, this.facebookLink, this.topDestinations,});
 
