@@ -20,26 +20,26 @@ class AllTripsRepository extends GenericBlocRepository<Trip> {
     List<Trip> _tripListFromSnapshot(QuerySnapshot snapshot) {
       try {
         List<Trip> trips = snapshot.docs.map((doc) {
-          Map<String, dynamic> data = doc.data();
+          Map<String, dynamic> data = doc.data() as Map<String, dynamic>;
           return Trip.fromData(data);
         }).toList();
         // trips.where((trip) => {
         //   trip.
         // })
         trips.sort(
-            (a, b) => a.startDateTimeStamp.compareTo(b.startDateTimeStamp));
+            (a, b) => a.startDateTimeStamp!.compareTo(b.startDateTimeStamp!));
         trips = trips
             .where(
-                (trip) => !trip.accessUsers.contains(userService.currentUserID))
+                (trip) => !trip.accessUsers!.contains(userService.currentUserID))
             .toList()
             .where((trip) =>
-                trip.endDateTimeStamp.toDate().isAfter(DateTime.now()))
+                trip.endDateTimeStamp!.toDate().isAfter(DateTime.now()))
             .toList();
         return trips;
       } catch (e) {
         CloudFunction()
             .logError('Error retrieving all trip list:  ${e.toString()}');
-        return null;
+        return [];
       }
     }
 
