@@ -22,18 +22,18 @@ class CurrentTripRepository extends GenericBlocRepository<Trip> {
         final now = DateTime.now().toUtc();
         var past = DateTime(now.year, now.month, now.day - 2);
         List<Trip> trips = snapshot.docs.map((doc) {
-          Map<String, dynamic> data = doc.data();
+          Map<String, dynamic> data = doc.data() as Map<String, dynamic>;
           return Trip.fromData(data);
         }).toList();
         List<Trip> crewTrips = trips
             .where(
-                (trip) => trip.endDateTimeStamp.toDate().compareTo(past) == 1)
+                (trip) => trip.endDateTimeStamp!.toDate().compareTo(past) == 1)
             .toList();
         return crewTrips;
       } catch (e) {
         CloudFunction()
             .logError('Error retrieving current trip list:  ${e.toString()}');
-        return null;
+        return [];
       }
     }
 
