@@ -20,17 +20,17 @@ class FavoriteTripRepository extends GenericBlocRepository<Trip> {
     List<Trip> _tripListFromSnapshot(QuerySnapshot snapshot) {
       try {
         List<Trip> trips = snapshot.docs.map((doc) {
-          Map<String, dynamic> data = doc.data();
+          Map<String, dynamic> data = doc.data() as Map<String, dynamic>;
           return Trip.fromData(data);
         }).toList();
         trips.sort(
-            (a, b) => a.startDateTimeStamp.compareTo(b.startDateTimeStamp));
+            (a, b) => a.startDateTimeStamp!.compareTo(b.startDateTimeStamp!));
 
         return trips;
       } catch (e) {
         CloudFunction()
             .logError('Error retrieving favorites trip list:  ${e.toString()}');
-        return null;
+        return [];
       }
     }
 
